requirements.txt
anyio==4.4.0
APScheduler==3.6.3
cachetools==4.2.2
certifi==2024.7.4
charset-normalizer==2.0.12
et-xmlfile==1.1.0
flake8==3.9.2
flake8-docstrings==1.6.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
mccabe==0.6.1
numpy==2.0.1
openpyxl==3.1.5
pandas==2.2.2
pycodestyle==2.7.0
pydocstyle==6.3.0
pyflakes==2.3.1
python-dateutil==2.9.0.post0
python-dotenv==0.19.0
python-telegram-bot==13.7
pytz==2024.1
requests==2.26.0
setuptools==71.0.3
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
tornado==6.4.1
tzdata==2024.1
tzlocal==5.2
urllib3==1.26.19
